initial startup commit

+++ ClemBot.Api/ClemBot.Api.Data/Models/Channel.cs
@@ -4,11 +4,11 @@ namespace ClemBot.Api.Data.Models
 {
     public class Channel
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         public string Name { get; set; }
 
-        public int GuildId { get; set; }
+        public ulong GuildId { get; set; }
         public Guild Guild { get; set; }
 
         public List<Message> Messages { get; set; } = new();
