+++ ClemBot.Api/ClemBot.Api.Data/Models/ClaimsMapping.cs
@@ -4,11 +4,11 @@ namespace ClemBot.Api.Data.Models
 {
     public class ClaimsMapping
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         public Claims Claim { get; set; }
 
-        public int RoleId { get; set; }
+        public ulong RoleId { get; set; }
         public Role Role { get; set; }
     }
 }
