+++ ClemBot.Api/ClemBot.Api.Data/Models/CustomPrefix.cs
@@ -2,11 +2,11 @@ namespace ClemBot.Api.Data.Models
 {
     public class CustomPrefix
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         public string Prefix { get; set; }
 
-        public int GuildId { get; set; }
+        public ulong GuildId { get; set; }
         public Guild Guild { get; set; }
     }
 }
