+++ ClemBot.Api/ClemBot.Api.Data/Models/DesignatedChannelMapping.cs
@@ -4,11 +4,11 @@ namespace ClemBot.Api.Data.Models
 {
     public class DesignatedChannelMapping
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         public DesignatedChannels Type { get; set; }
 
-        public int ChannelId { get; set; }
+        public ulong ChannelId { get; set; }
         public Channel Channel { get; set; }
     }
 }
