+++ ClemBot.Api/ClemBot.Api.Data/Models/Message.cs
@@ -4,19 +4,19 @@ namespace ClemBot.Api.Data.Models
 {
     public class Message
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         public string Content { get; set; }
 
         public DateTime Time { get; set; }
 
-        public int GuildId { get; set; }
+        public ulong GuildId { get; set; }
         public Guild Guild { get; set; }
 
-        public int ChannelId { get; set; }
+        public ulong ChannelId { get; set; }
         public Channel Channel { get; set; }
 
-        public int UserId { get; set; }
+        public ulong UserId { get; set; }
         public User User { get; set; }
     }
 }
