+++ ClemBot.Api/ClemBot.Api.Data/Models/Role.cs
@@ -2,13 +2,13 @@ namespace ClemBot.Api.Data.Models
 {
     public class Role
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         public string Name { get; set; }
 
         public bool? IsAssignable { get; set; }
 
-        public int GuildId { get; set; }
+        public ulong GuildId { get; set; }
         public Guild Guild { get; set; }
     }
 }
