+++ ClemBot.Api/ClemBot.Api.Data/Models/Tag.cs
@@ -5,7 +5,7 @@ namespace ClemBot.Api.Data.Models
 {
     public class Tag
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         public string Name { get; set; }
 
@@ -13,10 +13,10 @@ namespace ClemBot.Api.Data.Models
 
         public DateTime Time { get; set; }
 
-        public int GuildId { get; set; }
+        public ulong GuildId { get; set; }
         public Guild Guild { get; set; }
 
-        public int UserId { get; set; }
+        public ulong UserId { get; set; }
         public User User { get; set; }
 
         public List<TagUse> TagUses { get; set; }
