+++ ClemBot.Api/ClemBot.Api.Data/Models/TagUse.cs
@@ -4,17 +4,17 @@ namespace ClemBot.Api.Data.Models
 {
     public class TagUse
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         public DateTime Time { get; set; }
 
-        public int UserId { get; set; }
+        public ulong UserId { get; set; }
         public User User { get; set; }
 
-        public int TagId { get; set; }
+        public ulong TagId { get; set; }
         public Tag Tag { get; set; }
 
-        public int ChannelId { get; set; }
+        public ulong ChannelId { get; set; }
         public Channel Channel { get; set; }
     }
 }
